Split compiler directory
